Hide enrollment seat-holding internals behind EnrollmentService (#291)
CourseService previously injected EnrollmentRepository and referenced
EnrollmentStatus.SEAT_HOLDING_STATI directly, leaking enrollment's
capacity semantics across slices.

Move the seat-holder counting into EnrollmentService via three query
methods and a public RoleCounts record. CourseService now depends on
EnrollmentService and no longer knows which statuses count as
seat-holding. To avoid a CourseService ↔ EnrollmentService cycle,
EnrollmentService now injects CourseRepository for its tenant-scoped
course lookup, dropping the redundant findCourseByIdAndSchool helper.

As a side effect, EnrollmentRepository can return to package-private.
