+++ backend/src/main/java/ch/ruppen/danceschool/course/CourseRepository.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
-interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 
     boolean existsBySchoolId(Long schoolId);
